requirements.txt
numpy==1.26.4
sentence-transformers==2.4.0
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
python-dotenv==1.0.1
vec2text==0.0.12
fire==0.5.0
loguru==0.7.2
neptune==1.9.1
pytest==8.1.1
rouge-score==0.1.2
-e .